requirements.txt
google-generativeai>=0.8.0
streamlit>=1.38.0
python-dotenv>=1.0.0
requests>=2.31.0
APScheduler>=3.10.0
pydantic>=2.0.0
streamlit-cookies-controller>=0.0.3
pandas>=2.0.0
